Adding --runtimes to list installed runtimes
This switch to dotnet will list out the runtimes that are installed in the
"shared" folder. It will list out the install location (the location of
the "shared" folder) and all the folders that are underneath it.

The location is determined by using the muxer's location, that is, based
on the existing .NET Core SDK install conventions.