Further enhcanements of the projectless prototype
The logic is now moved to a separate class called ProjectlessSupport. A
sample F# project has been added and the dropping logic actually filters
out to the project file that is needed for the workspace in question.

More work is needed on detection of the workspace "type" (e.g. C# vs. F#
vs. VBNET).